vocab-importer/requirements.txt
python-fasthtml
python-dotenv
openai

